pkg/endpoint: move bpf-related code into bpf.go
The file endpoint.go is massive. Move BPF-related-code into bpf.go.

Signed-off by: Ian Vernon <ian@cilium.io>
