+++ pkg/endpoint/bpf.go
@@ -31,6 +31,7 @@ import (
 	"time"
 
 	"github.com/cilium/cilium/common"
+	"github.com/cilium/cilium/pkg/bpf"
 	"github.com/cilium/cilium/pkg/byteorder"
 	"github.com/cilium/cilium/pkg/completion"
 	"github.com/cilium/cilium/pkg/geneve"
@@ -49,6 +50,11 @@ import (
 const (
 	// ExecTimeout is the execution timeout to use in join_ep.sh executions
 	ExecTimeout = 300 * time.Second
+
+	// CallsMapName specifies the base prefix for EP specific call map.
+	CallsMapName = "cilium_calls_"
+	// PolicyGlobalMapName specifies the global tail call map for EP handle_policy() lookup.
+	PolicyGlobalMapName = "cilium_policy"
 )
 
 // lookupRedirectPortBE returns the redirect L4 proxy port for the given L4
@@ -820,4 +826,126 @@ func (e *Endpoint) regenerateBPF(owner Owner, epdir, reason string) (uint64, err
 		log.WithField(logfields.EndpointID, e.ID).WithError(err).Error("Exposing new bpf failed!")
 	}
 	return epInfoCache.revision, err
+}
+
+func (e *Endpoint) RemoveFromGlobalPolicyMap() error {
+	gpm, err := policymap.OpenGlobalMap(e.PolicyGlobalMapPathLocked())
+	if err == nil {
+		// We need to remove ourselves from global map, so that
+		// resources (prog/map reference counts) can be released.
+		gpm.DeleteIdentity(uint32(e.ID), policymap.Ingress)
+		gpm.DeleteIdentity(uint32(e.ID), policymap.Egress)
+		gpm.Close()
+	}
+
+	return err
+}
+
+// GetBPFKeys returns all keys which should represent this endpoint in the BPF
+// endpoints map
+func (e *Endpoint) GetBPFKeys() []lxcmap.EndpointKey {
+	key := lxcmap.NewEndpointKey(e.IPv6.IP())
+
+	if e.IPv4 != nil {
+		key4 := lxcmap.NewEndpointKey(e.IPv4.IP())
+		return []lxcmap.EndpointKey{key, key4}
+	}
+
+	return []lxcmap.EndpointKey{key}
+}
+
+// GetBPFValue returns the value which should represent this endpoint in the
+// BPF endpoints map
+func (e *Endpoint) GetBPFValue() (*lxcmap.EndpointInfo, error) {
+	mac, err := e.LXCMAC.Uint64()
+	if err != nil {
+		return nil, err
+	}
+
+	nodeMAC, err := e.NodeMAC.Uint64()
+	if err != nil {
+		return nil, err
+	}
+
+	info := &lxcmap.EndpointInfo{
+		IfIndex: uint32(e.IfIndex),
+		// Store security identity in network byte order so it can be
+		// written into the packet without an additional byte order
+		// conversion.
+		SecLabelID: byteorder.HostToNetwork(uint16(e.GetIdentity())).(uint16),
+		LxcID:      e.ID,
+		MAC:        lxcmap.MAC(mac),
+		NodeMAC:    lxcmap.MAC(nodeMAC),
+	}
+
+	for i, pM := range e.PortMap {
+		info.PortMap[i] = lxcmap.PortMap{
+			From: byteorder.HostToNetwork(pM.From).(uint16),
+			To:   byteorder.HostToNetwork(pM.To).(uint16),
+		}
+	}
+
+	return info, nil
+}
+
+// mapPath returns the path to a map for endpoint ID.
+func mapPath(mapname string, id int) string {
+	return bpf.MapPath(mapname + strconv.Itoa(id))
+}
+
+// PolicyMapPathLocked returns the path to the policy map of endpoint.
+func (e *Endpoint) PolicyMapPathLocked() string {
+	return mapPath(policymap.MapName, int(e.ID))
+}
+
+// IPv6IngressMapPathLocked returns the path to policy map of endpoint.
+func (e *Endpoint) IPv6IngressMapPathLocked() string {
+	return mapPath(cidrmap.MapName+"ingress6_", int(e.ID))
+}
+
+// IPv6EgressMapPathLocked returns the path to policy map of endpoint.
+func (e *Endpoint) IPv6EgressMapPathLocked() string {
+	return mapPath(cidrmap.MapName+"egress6_", int(e.ID))
+}
+
+// IPv4IngressMapPathLocked returns the path to policy map of endpoint.
+func (e *Endpoint) IPv4IngressMapPathLocked() string {
+	return mapPath(cidrmap.MapName+"ingress4_", int(e.ID))
+}
+
+// IPv4EgressMapPathLocked returns the path to policy map of endpoint.
+func (e *Endpoint) IPv4EgressMapPathLocked() string {
+	return mapPath(cidrmap.MapName+"egress4_", int(e.ID))
+}
+
+// PolicyGlobalMapPathLocked returns the path to the global policy map.
+func (e *Endpoint) PolicyGlobalMapPathLocked() string {
+	return bpf.MapPath(PolicyGlobalMapName)
+}
+
+func CallsMapPath(id int) string {
+	return bpf.MapPath(CallsMapName + strconv.Itoa(id))
+}
+
+// CallsMapPathLocked returns the path to cilium tail calls map of an endpoint.
+func (e *Endpoint) CallsMapPathLocked() string {
+	return CallsMapPath(int(e.ID))
+}
+
+// Ct6MapPath returns the path to IPv6 connection tracking map of endpoint.
+func Ct6MapPath(id int) string {
+	return bpf.MapPath(ctmap.MapName6 + strconv.Itoa(id))
+}
+
+func (e *Endpoint) Ct6MapPathLocked() string {
+	return Ct6MapPath(int(e.ID))
+}
+
+// Ct4MapPath returns the path to IPv4 connection tracking map of endpoint.
+func Ct4MapPath(id int) string {
+	return bpf.MapPath(ctmap.MapName4 + strconv.Itoa(id))
+}
+
+func (e *Endpoint) Ct4MapPathLocked() string {
+	return Ct4MapPath(int(e.ID))
 }
